The final swagger.yml and the built files that go with it.

+++ models/error_response.go
@@ -18,7 +18,7 @@ import (
 type ErrorResponse struct {
 
 	// code
-	Code int64 `json:"code,omitempty"`
+	Code int32 `json:"code,omitempty"`
 
 	// message
 	// Required: true
